Add a feature flag to revert retry handling
In case of issues found during our meetup, this PR adds a feature flag that we can use to go back to the old way of handling refunds.

N.B. ServiceLocator and feature flags are unavailable in the Hardware target, so this flag does not remove the changes to StripeCardReaderService.processPayment. This _should_ have minimal impact on our ability to turn off the key changes to payment retries.