Feature flag refreshing the order before payment
In case of issues found during our meetup, this PR adds a feature flag that we can use to go back to the old way of handling IPP orders – i.e. not refreshing them in the payment flow.

N.B. it’s not a perfect flag – it just short circuits the check method by calling the completion handler with `success` immediately, but it works pretty well.