Allow payment retry when there’s no PI to cancel
Payments which fail before the payment intent is created can still be retried.

They should follow the restart flow, which requires that the active payment intent be cancelled, and then a new PI be created for the retry.

Previously, if cancelling the payment intent failed because there was no active payment intent, we prevented the retry – which didn’t make sense.

If there’s no payment intent, it’s safe to make a new one. This commit adds an underlying error to that effect, and allows the retry to continue when this error is returned from a cancellation attempt.