10872 Check for iOS 16.4 to show TTP option in GB
We use Stripe’s `supportsReadersOfType` function to check whether Tap to Pay is available on a particular device. This includes a check for the supported version of iOS.

As of StripeTerminal 2.23.0, the minimum supported iOS version for TTP depends on the country in question. Unfortunately, `supportsReadersOfType` is not country-aware.

To avoid merchants in the UK with iOS 16-16.3.1 seeing a Tap to Pay on iPhone option which doesn’t work, we’ve added a check that they have at least iOS 16.4 in the StripeCardReaderService. US merchants continue to be able to use 16.0 and above.