10866 Rename Tap to Pay row after first transaction
We previously showed the `Set up Tap to Pay on iPhone` row with that title even after set up was complete. This was because it’s not possible for us to query the phone or Stripe for whether set up has been completed.

Set up needs to be completed for each store/device that’s used with TTP.

This commit adds a check for whether a TTP transaction has taken place using this site and phone, and if it has, changes the title to `Try out Tap to Pay`.

Users still may see a `Set up` screen when they tap on it, if they’re not connected to the built in reader. Most users will be connected though, due to TTP autoconnection on foreground.

This Set up screen simply connects to the reader and then leads them in to the Try out flow.

By making this change, we avoid making people on the Payments Menu think that they’ve not successfully set up Tap to Pay.