Configure UI tests to always store screenshots
Previously, the UI tests would deleted the screenshots in case of a
successful run. I feel that's a reasonable default because it does not
waste space. However, storage is cheap and we can build tooling to
garbage collect if necessary and I just run an investigation in which it
would have been useful to observe the recording for a successful test.
It's conceivable to imagine we'll run in similar situations in the
future and I believe we won't regret storing these assets.