Avoid race condition by using baseSiteURL when checking site existence
validateForm() updates loginFields.siteAddress to have appropriate scheme.

However, in between checkSiteExistence() and guessXMLRPCURL() async methods are called, loginFields.siteAddress can be changed by textField delegate methods back to an URL without a scheme.

This is a small change to make sure guessXMLRPCURL is called with an url with a scheme