Use type properties for Tracks thread safety
This is a somewhat speculative fix, which uses type properties to ensure thread safety of the creation of the TracksService, and its underlying Core Data store.

A lazy var, as in the previous implementation, is not guaranteed to be thread safe, and it was possible to see crashes in the app when two events were tracked concurrently on different threads.

See p91TBi-aJD-p2 for more details.