11009 Hide completed Set Up on About Tap to Pay
The About Tap to Pay screen can show a button for `Set up Tap to Pay on iPhone`, and does so by default.

Once we _know_ set up has been completed, that button can be confusing to users as it implies that it hasn’t been successful.

In the designs, we specified that it should be hidden on the About Tap to Pay screen, after it’s been completed.

As with the changes to the title of the `Set up/Try out` button, we rely on the heuristic of “have they completed a test payment” for this condition – if they complete set up but don’t try a payment, the button will remain.