11050 Fix reconnection after previous try fails
Previously, if an automatic reconnection to the Tap to Pay reader failed when the app was brought in to the foreground, the app would never try to reconnect again on subsequent foreground attempts, until it was either force quit or kicked from memory.

This resulted in merchants having to wait longer to be able to take a payment when they opened the app.

This commit changes the connection controller to allow reuse, by adopting the completion handler of a subsequent call to `searchAndConnect` without resetting its connection attempt, which would usually fail. When a connection attempt fails, it resets to the idle state, to allow it to be reused.

Additionally, the reconnection controller now reuses its connection controller when it still exists from a previous automatic reconnection attempt – which only happens the previous attempt failed.