11052 Fix auto-reconnect TTP on fresh launch
Previously, the auto-reconnection to the Tap to Pay system was not working on first launch; no connection was attempted. This reconnection is intended to save merchants time when taking a payment with TTP, by avoiding the wait for connection.

It was failing because in `reconnectIfNeeded` it checks for `case .completed = onboardingCache.value` in the guard. On first launch, this was always `nil`, because it hadn’t finished refreshing. Both the refresh and the reconnection are triggered by app launch and app foreground.

In this commit, we wait until the onboarding cache has finished refreshing, before we auto-reconnect to the Tap to Pay system.