11104 Prevent duplicate viewmodel creation
Previously, each time we opened the Payments menu, another InPersonPaymentsViewModel (i.e. for onboarding) would be created, resulting in many checks for onboarding.

This was due to using computed vars for child view model creation. The PaymentsMenu is long lived as it’s held by the HubMenu, and each time you open it, it’s redrawn. Each time the view’s redrawn, it was making a new set of child view models for all its child views.