11123 Use a UIVCRepresentable to fix VC dependency
The collect payment flows for IPP rely on having UIViewControllers passed to them to present the card reader connection alerts, and the onboarding views.

Breaking that dependency is out of scope of the Payments Menu rewrite, so we need a way to pass a UIViewController to these views.

The most effective way is to use the UIHostingController that we used to use, but wrap it up in a UIViewControllerRepresentable to present from SwiftUI.

I don’t like wrapping a wrapper… but that’s a problem for another day!