11132 Refresh Pay in Person state on Payments menu
Previously, changes to the Pay in Person toggle wouldn’t be reflected in the app until the store was switched, or the app was relaunched.

This PR adds a refresh call whenever the menu appears, which means that any changes from outside the app are reflected in the Payments Menu the next time it’s opened.