11131 Allow dismissal of webview from onboarding
A11y changes in a previous PR broke the Learn More webview dismissal when presented through SwiftUI wrappers.

The `task` that previously checked for a URL to present repeatedly re-presented the web view after it was closed, as the url couldn’t be set to nil

This commit retains the A11y behaviour by using a straightforward call to the customOpenURL handler from both actions.