11219 Use ScrollViewSectionKit in PaymentsMenu
SwiftUI Lists don’t allow different styling for different sections in the list.

The design for Deposits has the Deposit Summary in a full width view at the top of the payments menu list, while the rest of the payments menu is `insetGrouped`.

We can achieve this in a scrollview, but it’s pretty fiddly. Adding the `ScrollViewSectionKit` gives us lots of options for this styling.