11224 Show collapsed Deposits until expansion
The Deposits Summary is fairly large, and most of the time would require a user to scroll to do whatever they came to the Payments menu for.

The key information is the available and pending fund totals.

This commit collapses the view when the Payments Menu is loaded, and allows it to be expanded with a tap.