11168 restore toggle refresh test
Previously, the Payments Menu View Model directly refreshed the payment gateways.

From 03835cd, and the associated pr [#7633](https://github.com/woocommerce/woocommerce-ios/issues/7633),  we can see that this was to ensure that the state for the Pay in Person toggle was up to date.

We’ve moved the responsibility for actually refreshing the payment gateways to the toggle row view model… but `refreshState` is still called by the Payments Menu.

This restores the test, and updates it to check that we call `refreshState`.

We should probably move this responsibility to the `onAppear` of the toggle row though.