The major change is the introduction of a new method `addViewBelowStats`. This method is designed to add a view below the statistics view instead of the onboarding card.
The method first finds the index of the `headerStackView` in `containerStackView`. Then, it checks if `storeStatsAndTopPerformersViewController.view` exists and if it does, it finds its index in the `containerStackView`. If it doesn't exist, it defaults to the index of `headerStackView`. The new content view is then inserted after the index of the statistics view.

The `showBlazeCampaignView` method has been updated to use this new `addViewBelowStats` method instead of `addViewBelowHeaderStackView`. This ensures that the Blaze Campaign View is added below the statistics view.