Show Manage Card Reader modally
Previously, with UIKit, this was pushed with the tab bar hidden. This prevented changing tab and using the rest of the app with the connection controller active (and auto-disconnecting any tap to pay connections.)

Hiding the tab bar is not possible from SwiftUI on iOS 15, so instead we can present it as a sheet, like the Tap to Pay interface, which blocks the tab bar too.