11300 Show selected tab when long list available
Previously we did not update the scroll position for the visible tab headers when swiping currencies. Especially when the view was collapsed, this caused inconsistencies in the view and made it difficult to see which was selected.

This commit ensures that the tabs headers stay in sync with the visible content, and we scroll to the visible tab whenever they’re shown or hidden.