11334 Prevent ∞ loop in scrollVerticallyIfNeeded
Previously, we checked the size of the content within the body of the scrollVerticallyIfNeeded view modifier, and set the `shouldScroll` state according to the size compared against the parent’s size.

Since it was in `body`, this comparison happened during the layout pass, and because it resulted in a change of state which the same view depended on, it would result in a re-render of the view.

For some sizes, particularly where `contentSize` was very close/equal to `parentGeometry.size`, this re-render would change the `shouldScroll` state back to its previous value, and thus invalidate the view again. That was the cause of the infinite layout loop.

In this commit, we move the comparison out of the `body`, so we’re never changing `State` _during_ a layout pass, but instead we change it in response to a preference key change. `PreferenceKey` is a convenient way to pass data from child views to parent views.