11342 Restore Spotlight search for Payments
Before the SwiftUI payments menu rewrite, we used UIKit hooks to allow searchability of the Payments menu. Our registration code set the user activity as the ViewController’s user activity, which meant that the system handled `becomeCurrent` calls as appropriate.

Our workaround was to make a UIViewController that we only used for registering a user activity; however, since that was never displayed, it never became current, and the user activity wasn’t indexed.

This commit changes the workaround to directly call `becomeCurrent` on the user activity, so that it’s registered for search.

Additionally, we wait for the hub menu to be visible before presenting payments, to avoid it sometimes only getting as far as the hub menu.