11344 Show selected in manage payment gateways
Previously we forced a choice in the Payment Provider screen, without showing what they already had selected.

This wasn’t very clear. Furthermore we didn’t update the row on the payments menu to reflect their new choice until the onboarding was successful for that gateway, which made it look like we hadn’t made the change.

This commit updates the Payment Provider screen to show the currently selected plugin, where there is one, and updates the row according to the selected plugin even when onboarding is incomplete.