11649 Prevent crash by using column methods
The OrdersSplitViewWrapperController uses the UISplitViewController.Column API for the collapsing behaviour call (though not consistently through the rest of the class.) This collapsing behaviour applies on the phone.

In the `OrdersRootViewController`, where we handle showing the detail view, we used the older `showDetailViewController` API. Mixing and matching these approaches causes a crash on the phone, when you pass a UINavigationController subclass as the detail VC.

Split view controllers wrap plain VCs in a navigation view controller, but relying on this results in undesired behaviour; the back button on the phone takes you to the empty state, rather than the order list.