11466 Enable Payment Flow modals from SwiftUI
Previously, it was not possible, or at least clunky, to present the card payment flows modally over a modal presentation of the Payment Methods view from SwiftUI.

This was due to the use of the nearest navigationController as the root view for presenting card payment flow UI.

In this commit, we change to use the payment methods hosting controller to present card flows, not it’s nearest navigation controller, and we bring the nav into SwiftUI as well. This means that we present as follows:

OrderDetailsVC —modal—> PaymentMethodsHC —modal—> Payment flow.

Without this change, if the PaymentMethodsHC was in a modal, the navigation view of the OrderDetailsVC would be used to present both modals, which isn’t possible.