11466 Disable Collect Payment on empty orders
The Order Form can create empty orders, and it is possible to mark them as paid using some payment methods, but it’s not really a primary use case. It makes sense to disable the Collect Payment button until the merchant adds something to make the order have a non-zero total. The `Create` button is still available in the nav bar for this case.