11467 Improve view content changed height updates
Previously, when a change in data caused the expandable content view’s height to update, e.g. when adding or removing a shipping line on the OrderForm, the calculation was a little “off”. It would update the height, but not quite show enough.

This appears to be because the panel’s height was calculated before the view finished redrawing at its new height.

Dispatching to a later run cycle resolves the issue, similar to the way we handle these changes from tapping the expansion button.