11467 Vertical compact support for totals panel
In landscape on a phone, the totals panel on the OrderForm didn’t work well; its background/shadow didn’t extend into the horizontal safe areas, and if it ran out of space it just covered all the content.

This commit improves the situation by adding the panel as a safeAreaInset view on the scrollview, which is more like being in a ZStack than the previous VStack, which caused issues with capping the height of the panel at half the screen in some instances.

The panel is now capped at 80% the height of the screen, which is naïve but effective. When it hits the cap, the content becomes scrollable.