11483 Stop treating discovery success as an error
As of SDK version 3, the discovery completion handler behaviour has changed.

Previously, we treated any `nil` property for `recievedError` in this completion handler as an indication that discovery had been canceled. Since the SDK migration, we’ve been look for the specific SCPErrorCanceled to check that.

However, we also started to treat `nil` as an error, as we weren’t expecting it to happen, and we were a little confused by discrepancies in Stripe’s documentation, which said that it indicated cancelation in one place.

This quote appears to be key:
> discoverReaders is now completed successfully when connectReader is called.
– https://stripe.com/docs/terminal/references/sdk-migration-guide#update-your-discoverreaders-and-connectreader-usage

i.e. the completion handler is called. And there’s no error. So the property is nil.

I’ve confirmed that we get the call every time `connect` is called on a card reader, and previously that meant momentarily showing an error.

This commit removes that incorrect handling of `nil` and instead does nothing for discovery completion. The momentary errors no longer show during connection