11650 Prevent line deletion via steppers
Up to now, it’s been possible to remove a product by tapping the delete button in the product stepper. We previously made this easier to identify by changing the button to a red `x`, however it was still easy to accidentally trigger, and annoying to recover from (as the product needs to be found and added again.)

As one attempt to improve this, we tried adding an undo button in a notice after deletion, but it opened up too many issues of unreliability, e.g. sometimes adding the same product as two lines of the order.

Since there’s a dedicated `remove product from order` button, visible alongside the stepper, we’ve decided to remove this functionality from the stepper.

This commit disables the `-` button of the stepper when it’s at the minimum quantity.