11649 Only select first order when uncollapsed
A number of presentations on the iPad are horizontally compact, where we should behave as we do on the phone.

Previously we would always select the first order if running on an iPad, the first time you opened the order tab, regardless of whether it should show in the split view or not.

Additionally, we would sometimes switch to the first order on the phone as well, because of the call to `checkSelectedItem` in viewDidLoad which did not have any guard around it to prevent it happening on an iPad.

This commit makes selecting the first order the sole responsibility of the Order List (not the root view), and does it on viewWillAppear if we’re uncollapsed.

We will also need to do this when the size class changes, but that’s not essential for now.