11649 Use container VC, not a NavVC, on Orders tab
`UISplitViewController` is not supposed to be displayed within a `UINavigationController`, and when it is, it can lead to crashes during navigation, especially in horizontally compact size classes when the split view is collapsed.

Our `MainTabBarController` is set up with a `WooTabNavigationController` as the root of each tab, and when the store is known, we create the root content view controller for each tab and set it in the appropriate tab navigation controller’s `viewControllers` property (see `MainTabBarController.updateViewControllers(siteID:)`)

Since we can’t put our `OrdersSplitViewWrapperController` in a `UINavigationController` subclass without crashes, we instead need to wrap it in a plain `UIViewController`, so that we can retain the same `updateViewControllers` behaviour.

This commit adds a generic `WooTabContainerController` for this purpose, and sets it as the root of the Orders tab.