11649 Select Orders container type based on flag
Due to a longstanding iOS bug, there’s no obvious way to work around a black bar which shows at the bottom of the order list when it’s presented in a navigation controller, nested in the new `ordersContainerController`.

Since this is a temporary situation until we remove the feature flag, I’ve opted to set the tabs up differently depending on whether the flag is enabled or not.