11676 Show Collect Payment from Order Details
Using the new `onCompletion` block, we can wait until the Order Details screen is presented to show collect payment, when that’s tapped on the Order Form.

This involved finding the Order Details view, which handles both UISplitView based Orders tab, and the legacy orders tab. When the feature flag is removed, the guard blocks can be removed as well.