Only tap Back when the order split is collapsed
Now that the Orders tab is a split view, the Order List remains on screen with the Order Details screen (also known as SingleOrderScreen.)

This only happens when the app is presented with the horizontal size class `.regular`. (iPads mostly, but also landscape Plus/Pro Max iPhones).

If we call `pop()` in that scenario, the leftmost nav button in the Order List nav bar is tapped, which is the barcode scan button. The app gets stuck with that on screen.

This commit changes the `goBackToOrdersScreen` function to do no navigation unless it’s in `compact`.