11744 Always produce a ProductSelectorViewModel
There’s nothing which relies on `productSelectorViewModel == nil`, and with the existence of the `is…Presented` variable, no need to rely on that either.

Making sure we always have a `ProductSelectorViewModel` on hand makes it much easier to start off with the selector showing. We can still re-create it when the selector view is shown modally