11744 Use delegated presentation to show products
When we present the Order Form and Product Selector side-by-side on iPad, we don’t really want an Add Products button, because the products selector is already visible.

The AdaptiveContainerView provides a closure to show the secondary view when it’s not shown.

In this commit, we change to only show the add product button when the closure’s available (if the feature flag is on.)