11863 Call onCompletion with split view flag true
Previously we failed to call this `onCompletion` handler in this navigate method. We didn’t initially use it after tapping `Collect Payment`, however a change in f49f3e4 meant that we started to use this route.

Because we didn’t call `onCompletion`, the caller (`OrderDetailsRootViewController.setupNavigation`) didn’t get a chance to show the payment methods screen

This change makes sure we notify the caller of completion even when the feature flag is on, and fixes the presentation of the Payment Methods screen from Order Creation.