11887 Sync products immediately when side-by-side
With our previous a phone-centric layout, we relied on the merchant tapping a button to finish selecting products. The changes they made in the product selector weren’t synced until they tapped the button.

When the views are side by side, it doesn’t make sense to tap a button for your changes on one side of the screen to be reflected on the other.

In this commit, we make the presentation context available to the presented views, then use that to decide whether the button needs to be shown on the product selector, and if not, to sync the changes immediately.