11763 Deselect product when removed from OrderForm
The OrderForm shows all the product rows selected for the order. When expanded, there’s a `Remove item from order` button.

With the new side-by-side layout for OrderForm and ProductSelectorView, we need to update the product selector with changes from the OrderForm.

In this commit, we add a wrapper so that the entire OrderForm is a subview of the AdaptiveModalContainer, rather than just the product of a view builder being the order form side. This allows us to set a new `syncChangesImmediately` property on the view model, for use in deciding the correct behaviour when a product is removed.