11763 deselect on product selector when removed
When we remove an item from the order, using a button on the order form, we need to update the product selector’s state, if it’s visible.

`syncChangesImmediately` tells us whether we need to remove that, and if so, we remove the selected item from the product selector as well.