Load product selector configuration based on size class
The current `OrderFormPresentationWrapper` only checks for feature flag when loading the `AdaptiveModalContainer` or the `OrderForm`.

If the `AdaptiveModalContainer` is used, we always pass in the configuration for split views, which will cause issues when the presented view is not a split view (regular iPhone in portrait) since will use the configuration details specific for split views.

Since we render either ModalOnModalView or SideBySideView internally in AdaptiveModalContainer, we need to set a callback in order to notify the order form of changes, otherwise SwiftUI doesn’t seem to propagate changes when the size class changes.