11923 Use size class to set selected sync approach
The selected products are synced immediately when the product selector is shown in a split view, or on tapping a button when it’s shown as a modal.

This commit expresses these two approaches as an enum, and sets them based on the size class changing.

I’ve opted not to use the `adaptiveModalContainerPresentationStyle` (i.e. `modalOnModal` or `sideBySide`, as it wasn’t triggering `onChange(of: perform:)` as it should have, and is a plain wrapper for `horizontalSizeClass` anyway. We are planning to remove it later this week.