Update vertical spacing if order has products
The way we current render and switch different components was adding additional padding to the scanner button when opening an order in a tablet, if no products are added to the order.

This happens because no other subviews are rendered inside the ProductsSection when there’s no products. By calculating vertical spacing dynamically we resolve this issue.