11932 Fix selection update on back tap
Moving the `variationListViewModel` to the `ProductSelectorViewModel` (rather than a state variable on the `ProductSelectorView`) resulted in a lifecycle issue where tapping back would not store the updates.

`onDisappear` wasn’t a great place to put this call anyway, it would be called for other changes which aren’t related to the user tapping back.

Since we already add the back button ourselves, and disable swiping back, this is a straightforward fix for non-immediate sync contexts