12135 prevent legacy sheet showing with flag on
Previously, it was technically possible for the product selector sheet to be presented by the old code, over the new side-by-side view, when the feature flag is on.

In practice, it didn’t happen because we fixed `viewModel.isProductSelectorPresented` to false and handled it differently.

This commit makes it impossible for the old selector to show over the new view, to open up for future changes to the usage of `viewModel.isProductSelectorPresented`.