12135 Bind product selector presentation correctly
Previously, we kinda-sorta hardcoded `viewModel.isProductSelectorPresented` to true when the side by side view was showing (actually, we made it false and ignored it…)

This commit changes the `AdaptiveModalContainer` to correctly provide a binding, so that we can control the presentation in concert with the view or view model of the presented views.

The `AdaptiveModalController` reports changes to the presentation mode with its `task` modifiers, so when we shift from one presentation style to the other the flag gets correctly updated.