12135 Reset ProductSelectorViewModel when closed
When the product selector gets closed, we want to reset its view model.

If we don’t do that, changes are persisted from one presentation to the next, even though they’ve not been synced to the order… which is strange.

It’s doubly confusing because taps on products don’t _set_ the selection state, they _toggle_ it – so you get backwards behaviour.

This commit sets the viewmodel to nil when it gets closed, to achieve this.