12135 Use presentationStyle for Add Products visibility
Prior to this change, the Coupons product selector didn’t have an Add Products button when the feature flag was true.

Making the environment var optional allows us to be more accurate in consulting it; if the view isn’t embedded in an AdaptiveModalContainer, it will be `nil`.

Previous issues with `onChange` may still apply, but we don’t look for that notification here.