12170 Open order search results in split view
Previously we opened order search results modally.

With the new side-by-side view for order editing, it looked strange to open a new full screen modal over the formSheet search modal when shown on an iPad, when the merchant tapped edit from a search result.

This commit changes the search results to open in the split view context, when the side-by-side view is enabled.

Since they’re opening in the split view context, opening the side-by-side order editing screen above makes sense.