12170 Deselect on selection failure
Older orders are often not present in the datasource, when we try to select them.

If we have another row selected, and fail to select the new one, it is often confusing to keep that unrelated order selected, as we’re still showing the new order.

This change deselects the current selection in that situation, so at least we aren’t adding confusion for merchants.