dismiss modal on button tap if firmware update is completed
The cancel action closure is triggered when we tap the button on a firmware complete scenario.

Reusing this button through all modals also means that since we cannot cancel an update that has already been completed, it might fail to dismiss the modal when the cancellable logic is performed.

With this we allow the user to dismiss the modal manually if the automatic dismissal fails, unblocking the UI