Add entry to `en` `Localizable.strings` to unblock release
See failure at
https://buildkite.com/automattic/woocommerce-ios/builds/20356#018e3fea-e151-42a9-a088-b5d22ba90e1b
and conversation at p1710466850748509-slack-CC7L49W13

An addition to `Localizable.strings` landed on `trunk` after the code
freeze started but before it was merged, resulting in the new key being
uploaded to GlotPress but not being part of `Localizable.strings` in the
release branch.

The result was that the `strings` validation failed on the release
branch during deployment because of the unexpected entry.

To unblock the deployment, we decided to add the missing entry to the
release branch. The alternative would have been to delete every
localization from GlotPress, only for them to reappear on the next code
freeze upload, given the entry remains on `trunk`. It's safe to have
unused entries in `Localizable.string` so we chose this easier to
execute option.