12252 Prevent extra sync on open order creation
When we opened the create/edit order screen, we attempted to sync the order notes and customer details.

This was only supposed to happen on size class change – to avoid losing user entered data.

This commit changes it to only save when there’s a change.

We could still improve this by making the sync happen only when there’s a change from the synced value.