12257 Create datasource in viewDidLoad non-lazy
We know that we’ll need a datasource if viewDidLoad is called, so there’s no value in further delaying the init of the datasource.

`lazy var` comes with downsides, in particular that the init call is not thread safe.

We’ve seen crashes since the iPad split view order list support was added, particularly in
`OrderListViewController.configureViewModel`, which references `dataSource`. That, combined with other calls that the tableview could make, could be the cause of these crashes.

This commit moves the creation of dataSource to viewDidLoad, to prevent concurrency issues causing crashes.