12566 Prevent adding parent product via barcode
Scanning a barcode of a parent product previously added it to the order.

This isn’t possible in other ways, as normally a variation should be selected.

The implementation here shows the new order with a notice informing the user of the reason their product isn’t added. This matches Android.

A future improvement might be to open the product variation picker for the relevant product.