Add toTypeCheckedArray extension
The crash on peaMlT-vd-p2 indicates there could be unexpected data when converting from the NSOrderedSet [Storage.OrderItem] to a Swift array.

The new extension attempts to type check each element that will be passed into the array and assert in debug if something went wrong with the conversion.

Also updates the nil-coalescing fallback type in the read-only convertible implementation when fails to convert, so the empty array is explicitely of type `[Storage.OrderItem]` on compile time, rather than leaving it for type-inference on runtime.