Hook observation again after form reset
Observing published values for customer details only worked the first time we edit the form, but not after.

The issue comes from the form being deallocated upon dismissal. As we recreate the form on editing, at this point we’ve lost the reference that was set on EditableOrderViewModel init by `observeChangesInCustomerDetails()`, hence not observing new values that are published through the new object

This is fixed by calling again the observation hook when the reference is updated.