Fix notifications issues
In #12443, in-app notifications (when the app is in the foreground) were inadvertently broken, and so was navigation/handling of background notifications.

This was because we removed the delegate link to the UNUserNotificationCenter, so the delegate methods `AppDelegate.userNotificationCenter(_:didReceive:)` and `AppDelegate.userNotificationCenter(_:willPresent:)` were never called, and so not forwarded to the `pushNotesManager`.

This commit restores the delegate configuration.