Use long-lived connection controllers in the adaptor
At this point, I think the adaptor could be used as a singleton to give us one source of truth for card reader connections.

I’ve cut some corners in the POC and this still doesn’t totally prove that we can avoid UIKit use, but I _think_ the principle holds.