Refactor loading saved card steps to be more efficient.
Previously saved card loading is done whenever updateDashboardCards is called, which can be too frequent since that function s called by refreshDashboardCards that has a combine in it.
In this commit, the saved card is loaded once during init, then saved in `savedCards` variable. This variable is updated during user saving (didCustomizeDashboardCards), thus making it always be updated properly.