12867 Rename payment adaptor to service for clarity
The implementation of the CardPresentPaymentFacade happens to be one of the adaptors required to make the old payments code work… but that might not always be the case, and is an implementation detail that callers in the POS experience don’t need to be aware of.

Since this _should_ be in Yosemite or another similar level framework in the longer term, I’ve renamed it to Service for clarity at the call sites.