Fix duplicate ID warning in SwiftUI
When looping through a collection using keypaths the ID is suppossed to be unique hashables to differentiate views, but if we rely on the product ID this won’t happen since multiple of the same product ID can be added to the cart. We move to use their unique ID instead.