Basic dismissable alert
This adds a close button which actually works to the error modals shown… but it doesn’t remove the close button which is specified in the `CardPresentModalError` as the secondary button view model.

It’s a bit hacky, and might be improved by further adapting the viewModel instead of continuing to pass the `CardPresentPaymentModalViewModel` out to the swiftui which is presenting it… however, that’s easier said than done.