12868 Refactor showAlert to use event details
With this change, the `CardPresentPaymentService` starts to become more declarative.

We will (eventually) say “show this `eventDetails`”, and the caller will determine how that should be shown.

To make this clearer, I’ve renamed the alert view models to be POS-specific. Other callers in future would provide their own handling for a `CardPresentPaymentEvent`.

Leftover payment alert view models are untouched – these won’t be alerts anyway, so I’ll rename them in a future commit.

Note that this commit leaves payment messages not showing – this will be fixed in the next one.