Return card payment result when payment finishes
Previously, calls to `CardPresentPaymentCollectOrderPaymentUseCaseAdaptor().collectPaymentTask` would never return if the payment was successful, only if it was cancelled.

The completion handler we gave the `CollectOrderPaymentUseCase` wasn’t being called, because it was passed as the overall usecase’s completion. That closure is only called when the merchant dismisses or discards the receipt, in the existing modal, which marks the very end of the modal flow.

We should have actually used the `onPaymentCompletion` handler, which is called at the end of the `attemptPayment` completion handler, after a successful payment. Failed or cancelled payments return earlier, so we can safely say that any `onPaymentCompletion` call comes after a successful payment.

This commit fixes the issue by moving the contents of our adaptor’s `onCompleted` handler to the `onPaymentCompleted` handler instead. I also removed the `.idle` event, which had the result of preventing the `Payment successful` message from ever being shown.